requirements.txt
discord.py>=2.3.2,<3.0
yt-dlp>=2024.10.7
python-dotenv
aiohttp
aiosqlite
PyNaCl==1.5.0
